fix: onboarding atom

+++ src/features/onboarding/OnBoarding.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/native'
-import { useAtomValue, useSetAtom } from 'jotai/react'
-import React, { useEffect } from 'react'
+import { useAtom } from 'jotai/react'
+import React, { useEffect, useState } from 'react'
 import Modal from 'react-native-modal'
 import { useDispatch } from 'react-redux'
 import { deleteAllDatabases } from '~helpers/sqlite'
@@ -8,7 +8,7 @@ import useLanguage from '~helpers/useLanguage'
 
 import { getIfVersionNeedsDownload } from '~helpers/bibleVersions'
 import { setDefaultBibleVersion } from '~redux/modules/user'
-import { isOnboardingRequiredAtom } from './atom'
+import { isOnboardingCompletedAtom } from './atom'
 import DownloadResources from './DownloadResources'
 import OnBoardingSlides from './OnBoardingSlides'
 import SelectResources from './SelectResources'
@@ -24,39 +24,49 @@ const StylizedModal = styled(Modal)(({ theme }) => ({
 const useCheckMandatoryVersions = () => {
   const isFR = useLanguage()
   const dispatch = useDispatch()
-  const setIsOnboardingRequired = useSetAtom(isOnboardingRequiredAtom)
+  const [isOnboardingCompleted, setIsOnboardingCompleted] = useAtom(isOnboardingCompletedAtom)
+  const [showOnboarding, setShowOnboarding] = useState(false)
 
   useEffect(() => {
+    // Skip file check if onboarding was already completed (fast path via MMKV)
+    if (isOnboardingCompleted) {
+      console.log('[Onboarding] Already completed, skipping file check.')
+      return
+    }
+
     ;(async () => {
       try {
         const lsgNeedsDownload = await getIfVersionNeedsDownload(isFR ? 'LSG' : 'KJV')
 
         if (lsgNeedsDownload) {
           console.log('[Onboarding] Needs download, open onboarding.')
-          setIsOnboardingRequired(true)
+          setShowOnboarding(true)
           dispatch(setDefaultBibleVersion(isFR ? 'LSG' : 'KJV'))
           deleteAllDatabases()
         } else {
-          setIsOnboardingRequired(false)
+          // Bible exists, mark onboarding as completed for future fast starts
+          console.log('[Onboarding] Bible exists, marking as completed.')
+          setIsOnboardingCompleted(true)
         }
       } catch (error) {
         console.error('[Onboarding] Error checking version:', error)
         // On error, assume onboarding is required to be safe
-        setIsOnboardingRequired(true)
+        setShowOnboarding(true)
         dispatch(setDefaultBibleVersion(isFR ? 'LSG' : 'KJV'))
       }
     })()
-  }, [isFR, dispatch])
+  }, [isFR, dispatch, isOnboardingCompleted])
+
+  return showOnboarding
 }
 
 const OnBoarding = () => {
   const [step, setStep] = React.useState<number>(0)
-  const isOnboardingRequired = useAtomValue(isOnboardingRequiredAtom)
-  useCheckMandatoryVersions()
+  const showOnboarding = useCheckMandatoryVersions()
 
   return (
     // @ts-ignore
-    <StylizedModal isVisible={isOnboardingRequired === true}>
+    <StylizedModal isVisible={showOnboarding}>
       {step === 0 && <OnBoardingSlides setStep={setStep} />}
       {step === 1 && <SelectResources setStep={setStep} />}
       {step === 2 && <DownloadResources />}
